requirements.txt
pyyaml>=6.0
paho-mqtt>=1.6.1
pycron>=3.0.0
google-api-python-client>=2.37.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.4.6
